1.22 마이페이지 완료..? CSS 다듬고 중복체크 등등 남음...
