+++ BookJeok/src/main/webapp/js2/alert.js
@@ -67,4 +67,8 @@ function likeItCKModal() {
 	})
 } // end of likeItCKModal
 
+//문의내역(없음)
+
+
+
